Extract functions used by IDEA plugin to utility class
